requirements.txt
pynput==1.8.1
sounddevice==0.5.5
numpy==2.4.4
scipy==1.17.1
PyYAML==6.0.3
dashscope==1.25.15
